refactor: lower_inst 拆为 LowerCtx + 按语义分家族的 inst/ 模块
改前:lower_inst / _b / _c 三个文件是按行数切的,同一份 12 参数签名
(全带 #[allow(clippy::too_many_arguments)]),靠 lower_inst.rs:709 →
lower_inst_b → lower_inst_c 链式 fallthrough。后果是'新 opcode 该加在
哪'没有答案,只能往末尾堆。

- 新增 LowerCtx:可变的 ssa/globals/sig + 只读的 func/funcs/entry/
  module_globals/capture_params。每个 lowering 例程 12 参数 → 5 参数,
  三处 too_many_arguments allow 全部消失
- inst/mod.rs 的 match 成为唯一路由表(opcode → 语义家族);家族模块:
  scalar(常量/move/谓词/算术/比较)、string(字面量/display/拼接/
  split/join)、call(直接调用/方法分派/闭包/去虚化)、global(可变
  全局 + 捕获 cell)、container(list/map/object 构造·索引·变更)、
  control(Raise);未列入路由表的 opcode 落到 Unsupported——干净回退,
  不会误编译
- scalar.rs(值读取/coerce/display helper)改名 convert.rs,与新的
  inst/scalar.rs(标量 opcode lowering)区分
- lib.rs 的 crate doc 已严重过时(称 map/闭包/动态索引不支持,且提到
  已退役的 legacy 后端),改写为真实 scope + 模块地图;
  aot-redesign.md §4 的模块结构同步为落地形态

纯结构调整,arm 内代码逐字搬移。cargo test --workspace --all-features
109 个 test result 全绿(与改前一致);clippy 零警告;
aot_coverage 51/51 不变
